Update workflow platform configuration
When using the matrix strategy to define the arches to build
and push the image only one arch is being retained on Quay.
It seems the second build always overwrites the first build,
leaving only one arch on Quay. This change puts both arches
into the same build step to prevent this happening.